Move ancestry tests
